Decouple connection status from data sources
