+++ backend/app/schemas/common.py
@@ -1,8 +1,18 @@
-from typing import Generic, List, TypeVar
+from datetime import datetime
+from typing import Generic, List, Literal, Optional, TypeVar
 
 from pydantic import BaseModel
 
 from app.schemas.filter import BaseFilterModel
+
+
+class ConnectionStatus(BaseModel):
+    """Reusable connection test status for integrations"""
+
+    status: Literal["Untested", "Connected", "Error"] = "Untested"
+    last_tested_at: Optional[datetime] = None
+    message: Optional[str] = None
+
 
 T = TypeVar("T")
