+++ frontend/src/interfaces/connectionStatus.interface.ts
@@ -1,0 +1,5 @@
+export interface ConnectionStatus {
+  status: "Untested" | "Connected" | "Error";
+  last_tested_at: string | null;
+  message: string | null;
+}
