+++ frontend/src/views/DataSources/components/DataSourceDialog.tsx
@@ -25,17 +25,10 @@ import {
   SelectValue,
 } from "@/components/select";
 import { toast } from "react-hot-toast";
+import { Loader2 } from "lucide-react";
+import { ConnectionTestPanel } from "@/components/ConnectionTestPanel";
+import type { ConnectionStatus } from "@/interfaces/connectionStatus.interface";
 import {
-  AlertCircle,
-  CheckCircle,
-  HelpCircle,
-  Loader2,
-  Plug,
-} from "lucide-react";
-import { Alert, AlertDescription } from "@/components/alert";
-import { Badge } from "@/components/badge";
-import {
-  ConnectionStatus,
   DataSource,
   DataSourceConfig,
 } from "@/interfaces/dataSource.interface";
@@ -452,73 +445,11 @@ export function DataSourceDialog({
 
                 {/* Test connection */}
                 {!isOAuthType && (
-                  <div className="space-y-2">
-                    <div className={`flex justify-end${isTesting ? " invisible" : ""}`}>
-                      {testStatus?.status === "Connected" ? (
-                        <Badge
-                          variant="default"
-                          className="text-green-700 bg-green-100"
-                        >
-                          <CheckCircle className="w-3 h-3 mr-1" /> Connected
-                        </Badge>
-                      ) : testStatus?.status === "Error" ? (
-                        <Badge variant="destructive">
-                          <AlertCircle className="w-3 h-3 mr-1" /> Error
-                        </Badge>
-                      ) : (
-                        <Badge variant="outline">
-                          <HelpCircle className="w-3 h-3 mr-1" /> Untested
-                        </Badge>
-                      )}
-                    </div>
-                    {isTesting ? (
-                      <Alert>
-                        <Loader2 className="h-4 w-4 animate-spin" />
-                        <AlertDescription>
-                          Testing connection, please wait…
-                        </AlertDescription>
-                      </Alert>
-                    ) : testStatus?.status === "Connected" ? (
-                      <Alert variant="success">
-                        <CheckCircle className="h-4 w-4" />
-                        <AlertDescription>
-                          {testStatus.message}
-                        </AlertDescription>
-                      </Alert>
-                    ) : testStatus?.status === "Error" ? (
-                      <Alert variant="destructive">
-                        <AlertCircle className="h-4 w-4" />
-                        <AlertDescription>
-                          Connection failed. Please verify your settings.
-                          {testStatus.message && (
-                            <span className="block mt-1 text-xs opacity-75 break-all">
-                              {testStatus.message}
-                            </span>
-                          )}
-                        </AlertDescription>
-                      </Alert>
-                    ) : (
-                      <Alert>
-                        <AlertCircle className="h-4 w-4" />
-                        <AlertDescription>
-                          This configuration has not been tested. Verify it works before saving.
-                        </AlertDescription>
-                      </Alert>
-                    )}
-                    <Button
-                      type="button"
-                      className="w-full"
-                      onClick={handleTestConnection}
-                      disabled={isTesting}
-                    >
-                      {isTesting ? (
-                        <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                      ) : (
-                        <Plug className="mr-2 h-4 w-4" />
-                      )}
-                      Test Connection
-                    </Button>
-                  </div>
+                  <ConnectionTestPanel
+                    isTesting={isTesting}
+                    testStatus={testStatus}
+                    onTest={handleTestConnection}
+                  />
                 )}
               </>
             )}
